+++ CHANGELOG.md
@@ -58,6 +58,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added field utilities to perform basic numeric operations on fields
+
 ### Changed
 
 - Updated programs using FLAP for command line parsing to use fArgParse instead